+++ examples/policies/token_introspection_configuration.lua
@@ -1,0 +1,14 @@
+local policy_chain = require('apicast.policy_chain').default()
+
+local token_policy = require('apicast.policy.token_introspection').new({
+        introspection_url = "http://localhost:8080/auth/realms/3scale/protocol/openid-connect/token/introspect",
+        client_id = "YOUR_CLIENT_ID",
+        client_secret = "YOUR_CLIENT_SECRET"
+})
+
+policy_chain:insert(token_policy)
+
+return {
+        policy_chain = policy_chain
+}
+